--- a/include/ipsecconf/confread.h
+++ b/include/ipsecconf/confread.h
@@ -71,7 +71,6 @@ struct starter_end {
 	ip_address addr;
 	ip_address nexthop;
 	ip_cidr vti_ip;
-	ip_protoport protoport;
 
 	keyword_values values;
 };