Issue: #2002 

Scenario:
1. We have an `ipsec.conf` with a 1000 connections specified.
2. Connections have `rightprotoport=udp/6081` or `leftprotoport=udp/6081`
3. We add each of these connections with `ipsec add <conn>` one by one, since there are other connections we don't want to disrupt in the process.

Result:
- It takes about 30 minutes to complete the configuration.

Cause:
- Every `ipsec add <conn>` results in full parsing of the large `ipsec.conf`.
- While parsing `*protoopt=udp/6081` of every connection, a very heavy `getservbyname()` function is called (takes 1.8 ms per call).

This PR attempts to solve the problem by:
1. Only loading connections that are specified on the command line (as possible), to avoid unnecessary memory allocations and extra parsing of non-interesting connections.
2. Reordering the `protoport` parsing logic to first check if the value is a number to avoid the costly `getservbyname()` call.

Results:
1. 1000x `ipsec add <conn>`:
```
        protoport      main       Patch #1    Patch #2      Both     SpeedUp
        --------------------------------------------------------------------
        udp/6081      29.6 min    24.5 sec    54.4 sec    22.5 sec     79x
        udp/geneve    22.9 min    23.9 sec    22.9 min    23.9 sec     57x
```
2. `addconn --checkconfig` (only affected by the change in `ttoport`):
```
        protoport      Before     After     SpeedUp
        -------------------------------------------
        udp/6081      1.77 sec   0.04 sec     44x
        udp/geneve    1.37 sec   1.37 sec     --
```

Technically, either one of the changes solves the problem for the main use-case with Open vSwitch (`udp/6081`), but it seemed better to fix the issue for both.

Note: I'm not particularly happy with parsing of aliases and recursive checks for `also`, but I didn't find a better way to filter out connections without breaking the logic for aliases.

Testing done with the following simple script (adjust `N` to get +/-250 connections):
```
# cat pluto-test.sh 
kill $(cat /tmp/pluto-test/pluto.pid 2>/dev/null) 2>/dev/null
sleep 2
ip netns del pluto-ns 2>/dev/null
ip netns add pluto-ns
ip -netns pluto-ns link set dev lo up
ip link del pluto-p1 2>/dev/null
ip link add pluto-p1 type veth peer name pluto-p2
ip link set pluto-p2 netns pluto-ns
ip -netns pluto-ns link set dev pluto-p2 up
ip -netns pluto-ns addr add 30.30.1.1/16 dev pluto-p2
ip -netns pluto-ns route add default via 30.30.100.254

rm -rf /tmp/pluto-test
mkdir -p /tmp/pluto-test/ipsec.d
echo '
config setup
    uniqueids=yes

conn %default
    keyingtries=%forever
    type=transport
    auto=route
    ike=aes_gcm256-sha2_256
    esp=aes_gcm256
    ikev2=insist
' > /tmp/pluto-test/ipsec.conf

touch /tmp/pluto-test/secrets
ipsec initnss --nssdir /tmp/pluto-test/ipsec.d
ip netns exec pluto-ns ipsec pluto --config /tmp/pluto-test/ipsec.conf \
      --ipsecdir /tmp/pluto-test --nssdir /tmp/pluto-test/ipsec.d \
      --logfile /tmp/pluto-test/pluto.log \
      --secretsfile /tmp/pluto-test/secrets \
      --rundir /tmp/pluto-test

sleep 2

set -o errexit

N=5
M=251
for i in $(seq 2 ${N}); do
    for j in $(seq 2 ${M}); do
        echo "
conn tunnel-${i}-${j}
    left=30.30.1.1
    right=30.30.${i}.${j}
    authby=secret
    leftprotoport=udp
    rightprotoport=udp/6081
" >> /tmp/pluto-test/ipsec.conf
        echo "
%any 30.30.${i}.${j} : PSK \"swordfish\"
" >> /tmp/pluto-test/secrets
    done
done

ip netns exec pluto-ns ipsec \
    --ctlsocket /tmp/pluto-test/pluto.ctl \
    --config /tmp/pluto-test/ipsec.conf   \
    rereadsecrets >/dev/null 2>&1

start_time_total=$(date +%s.%3N)
k=1
for i in $(seq 2 ${N}); do
    for j in $(seq 2 ${M}); do
        start_time=$(date +%s.%3N)
        ip netns exec pluto-ns ipsec \
            --ctlsocket /tmp/pluto-test/pluto.ctl \
            --config /tmp/pluto-test/ipsec.conf   \
            add tunnel-${i}-${j} \
            >> /tmp/pluto-test/add.log 2>&1
        end_time=$(date +%s.%3N)
        add_s=$(echo "scale=3; ${end_time} - ${start_time}" | bc)
        printf "%-4d %5.3f\n" ${k} ${add_s}
        let k=$(($k + 1))
    done
done
end_time_total=$(date +%s.%3N)
total_s=$(echo "scale=3; ${end_time_total} - ${start_time_total}" | bc)
printf "Total: %5.3f seconds\n" ${total_s}

kill $(cat /tmp/pluto-test/pluto.pid)
sleep 2
rm -rf /tmp/pluto-test
ip link del pluto-p1
ip netns del pluto-ns
```