@igsilya 2025-01-15
> What about dropping `struct starter_end .protoport` and leaving the value as a string? Then `set_whack_end()` can do the conversion.

Wouldn't that break `addconn --checkconfig` functionality, i.e. it will not fully validate the config anymore?  Being able to validate the config without communicating with `pluto` also seem like a good thing.