@cagney 2025-01-15
> Wouldn't that break `addconn --checkconfig` functionality, i.e. it will not fully validate the config anymore?

That ship has long since departed.

Both `whack` and `addconn` leave most things to `pluto`.  That way there's only one block of code cross checking the config and applying defaults.  It also means that pluto can log what was attempted and rejected.  This is a big improvement on 4.x, where `whack`, `addconn` and `pluto` each implemented a haphazard subset of contradictory checks and defaults that had a habit of failing silently.

> Being able to validate the config without communicating with `pluto` also seem like a good thing.

True.  The file is still parsed so the basic syntax is still validated (it isn't the parser that is slow, its the verification that comes later).

One thought would be for  `addconn` and `whack` to send connections to `pluto` for additional verification (but with being adding).

Another would be to re-organize `addconn`, so that, when check-config, it goes through the motions of building each whack message but not send it.  It might then be possible to move some of the simpler checks performed by `pluto` to a library where they can be shared.