@igsilya 2025-01-15
> > Wouldn't that break `addconn --checkconfig` functionality, i.e. it will not fully validate the config anymore?
> 
> That ship has long since departed.

OK.  So I tried the suggested solution.  Results are as follows:

Patches:
1. [addconn: do not load connections that were not requested](https://github.com/libreswan/libreswan/pull/1987/commits/5ae40dc536f4a5b59165914dc18a70e7069ab824)
2. [ttoport: check numbers before the service name lookup](https://github.com/libreswan/libreswan/pull/1987/commits/8c14ad9e3c7eb9e07444d619538dbfe847713597)
3. ipsecconf: move protoport parsing to set_whack_end (NEW)

---

1. 1000x `ipsec add`
```
    protoport      main       Patch #1    Patch #2    Patch #3      1+2       2+3       1+2+3
    -------------------------------------------------------------------------------------------
    udp/6081      29.6 min    24.5 sec    54.4 sec    55.0 sec    22.5 sec   52.0 sec   22.4 sec
    udp/geneve    22.9 min    23.9 sec    22.9 min    55.0 sec    23.9 sec   54.0 sec   23.9 sec
```
2. `addconn --checkconfig`
```
    protoport      main       Patch #1    Patch #2    Patch #3      1+2       2+3       1+2+3
    --------------------------------------------------------------------------------------------
    udp/6081      1.77 sec    1.77 sec    0.04 sec    0.04 sec    0.04 sec   0.04 sec   0.04 sec
    udp/geneve    1.37 sec    1.37 sec    1.37 sec    0.04 sec    1.37 sec   0.04 sec   0.04 sec
```

It seems like not loading connections (patch 1) is still about 2x faster than `set_whack_end()` approach (patch 3) in the `ipsec add` case and the `set_whack_end()` approach wins the `checkconfig` contest.   Having all 3 patches seems benefitial.  However, we could drop some of them.  For example, if we keep only the  `set_whack_end()` patch, then all `ipsec add` calls can be done in under a minute, which is good, though not as good as 25-ish seconds, obviously.
I'd keep patch 2 (the `ttoport` reorder) regardless, since it seem to make sense on its own even if not very beneficial in a grand scheme of things if the other changes get accepted.

What do you think?