@cagney 2025-01-15
> 1. [addconn: do not load connections that were not requested](https://github.com/libreswan/libreswan/pull/1987/commits/5ae40dc536f4a5b59165914dc18a70e7069ab824)

See below ...

> 2. [ttoport: check numbers before the service name lookup](https://github.com/libreswan/libreswan/pull/1987/commits/8c14ad9e3c7eb9e07444d619538dbfe847713597)

I've merged this.

>  3. ipsecconf: move protoport parsing to set_whack_end (NEW)

Can you push this.  It should be good enough.

Presumably the discrepancy between 1. and 3. is all the calls to `validate_end()` expanding connections.   I suspect the work can be moved closer to the `starter_whack_add_conn()` call so it only happens when needed.  That should have the same effect as patch 1., but without the hoops.