@igsilya 2025-01-15
> I've merged this.

Thanks!

> > 3. ipsecconf: move protoport parsing to set_whack_end (NEW)
>
> Can you push this. It should be good enough.

I rebased the PR branch and added this patch: https://github.com/libreswan/libreswan/pull/1987/commits/cef92a7be691a9d948594274924cafea0550909f

> Presumably the discrepancy between 1. and 3. is all the calls to `validate_end()` expanding connections. I suspect the work can be moved closer to the `starter_whack_add_conn()` call so it only happens when needed. That should have the same effect as patch 1., but without the hoops.

Ack.  I'll try to do some profiling around this later.  But, I agree, that having patches 2 and 3 should be good enough for now.